HS-248/back/feature/word game/result/get 사용자는 낱말 게임이 끝나면 게임 참여자들의 점수와 등수를 최종적으로 볼 수 있다.

+++ backend/src/main/java/com/hans/hans/domain/wordgame/dto/WordGameResultResponseDto.java
@@ -9,5 +9,5 @@ import java.util.Map;
 @AllArgsConstructor
 public class WordGameResultResponseDto {
 
-    private Map<String,Integer> players;
+    private Map<String,Long> players;
 }
